# Return iterator in tokenizers